Inital commit of DogStatsd client
This is an implementation of DogStatsd client using QuickCheck style
property based testing, and a strong focus on simplicity. The driving
factors are to make a robust, easy to use, and easy to understand libary
for working with Datadog.

This commit starts us off with the core message processing and
associated property tests. In an effort to keep the code clean, and easy
to read, the knee jerk reaction to create a GenServer has been delayed.
The result is pure functions which can be easily tested. It is expected
we should be able to build a good immutable queue structure to support
buffering when the need arises, but for now, this is great.

The next step will be to introduce the ability to send the packets via
udp. This still does not require us to use GenServer.